Initial commit
This was originally hashicorp/terraform#7058, however due to issues
(documented in the README) it can't be included in Terraform mainline.
More than a few people have still found it useful though, and an
external plugin has been requested.

This has been tested as of today and is still functional (the HTTP test
was skipped but the Route53 tests are all functional, and the plugin
works outside of the tests).

Also added code to use the dnssimple and powerdns lego modules.