certificate: Trap DNS challenge initialization errors properly
setDNSChallenge was not catching errors, even though I had set up the
function to do that properly. This fixes it so that if there is an error
actually setting up the DNS challenge (ie: missing credentials or other
settings), TF will abort instead of just attempt a HTTP or TLS
challenge.

Related: #6