Vendor "fun"
Not too sure why, but govendor fetch is not fetching from my origin
correctly (so it can't grab the updated fork for ResourceDiff properly).
Had to add manually with govendor add, but of course, that added a bunch
of other junk.

The updated copy in Terraform here now has the custom diff fork as per
before, but now rebased against v0.10.0-beta2. The PR will be tracking
releases now heading up to full v0.10.0 release.