Remove "experimental" from CHANGELOG
"Experimental" is not necessarily the right word here. While the PR is
still pending and may undergo changes, the large majority of the changes
that the custom diff patch entail are behind the plugin boundary, and
ResourceProvider interface, without any change to the binary plugin RPC
data exchanged. This is perfectly legit in the the current provider
behaviour specification, as helper/schema, while being pretty much the
de facto resource API implementation, is still just an implementation
and is open to modification by a provider owner so long as they respect
the plugin protocol.