vendor: Lock copystructure and reflectwalk to TF-locked versions
There was a bug in copystructure that was causing pointer types to not
be copied properly (instead it would just return the value as the type
it was pointing to). Unfortunately, this bad behaviour propagated to
CustomizeDiff as the incorrect behaviour was tested against and locked
in, ultimately. TF is still vendored against the buggy version.

Until this can be fixed, we have to override copystructure (and
reflectwalk just to be safe) to the versions that they are vendored as
in Terraform on the current master.

I will probably work on fixing this in TF core soon, but for now this
ensures things continue to work.