Update tests for certificate resource
This updates the tests for the certificate resource, adding the new zone
ID environment variable and fixing the direct option DNS provider test
so that it clears the environment before attempting the test, ensuring
that options come from config. The variables are re-set to their
original values after the test.

It looks like the renewal test is failing due to a non-empty plan, but
I'm wondering if that's a post-0.4.0 thing - since the certificate is
now being properly forced to NewComputed in the plan when it has
expired, there is always going to be a diff unless we do something like
create another step with the renewal time set back to the default.