structure: Remove expandACMERegistration
This function has not been used in quite a long time, is broken in the
new lego API, and is actually superfluous in the new API as there are no
fields that are currently expanded on in this function that aren't by
expandACMEClient.

We are able to delete registrations now, but all of the items necessary
to do that are currently populated by expandACMEClient, so that will be
used when we want to delete registrations.