[WIP] Fundamental client building/registration changes
Due to the changes in both ACME spec and the lego API, an ACME
registration/account is now a very small resource, consisting of mainly
only the ID. All other fields are ultimately superfluous - registration
body is a field, but is currently not exported, and registration URL
will change from the transition from ACME v1 to v2, more than likely
causing issues with existing certificates as the URL change would force
a new resource.

As a result, we are dropping all unique fields from acme_registration,
save email_address.

This has introduced fundamental changes to how we build the ACME client,
which are reflected in this WIP. This commit also has most of the
changes to the acme_registration resource that reflect its final state.
Final changes need to be made to the rest of the code to make sure
things build again.