r/certificate: Remove OCSP revocation validation
Back when the ACME provider was written in 2016, the Let's Encrypt OCSP
endpoint was (seemingly) a lot more stable than it has been as of late.
We have had bug reports and issues with both manual and automatic
testing all surrounding issues with timeouts on the OCSP endpoint on
destroy.

This is not necessarily a problem with just Let's Encrypt - OCSP has
historically been unreliable, with several countermeasures being taken
by a number of webservers and web browsers to try and combat outages
that can sometimes last days. Some applications, like Chrome, don't even
use OCSP, opting for their own CRL lookup technology (CRLSets).

Given that this has been more of a hindrance than a help lately, I've
opted to remove revocation validation from the provider. Deletion is now
a fire and forget operation, and if it succeeds on the ACME endpoint,
we assume the CA will do the job of revocation as it has been asked to.

Fixes #41.