r/certifiate: ID should change on renewal
In #48 we restored the URL attributes to both resources and ensured that
ID modification was de-coupled from the resource lifecycle aside from
create.

This is not technically correct for acme_certificate - as when a renewal
is done with lego, it is an entirely new certificate. I don't know if
this is an ACME or lego shortcoming, but it doesn't necessarily matter
as we mainly care about how it's currently implemented in the client.

This updates things so that we call a SetId on renewal with the new
certificate's URL.

Fixes #52.