docs: Start creating website documentation for upstream release
This is the start of the documentation for the upstream release. As part
of this, I am also documenting the DNS providers, that users do not have
to refer back to the lego docs anymore.

Note that some of these docs will refer to links that do not exist yet.