website: Revert legacy provider location back to vancluever
After some discussion it's been decided that we are not well set up for
a scenario where the repo is transferred versus just duplicated, as is
our normal procedure.

We will just duplicate the repo and I will copy over any existing issues
by hand when we do the release, which will more than likely be tomorrow.