modules: Update lego to v1.1.0, egoscale to v0.9.28
There were some build issues with egoscale v0.9.27 (missing
UpdateDNSRecord) type, which is what modules pulled in after we
updated to v1.1.0. These were fixed with a version bump.

The basic certificate test works, so I'm going to merge the changes
here so that we can do the rest of the work necessary to get lego to
v2.