Post-rebase go mod cleanup, ensure CADirURL is set
Most of the noise here is go module cleanup after the rebase.

One small fix has been made to the changes made to work with the new
API - namely, CADirURL was not set and as a result all connections
were defaulting to the live Let's Encrypt environment. This was
discovered after consecutive acceptance tests ultimately caused a
registration limit.