Code cleanup and more post-update stuff
Some more post-update stuff:

* Reverted a nil check on a user registration in expandACMEClient
that will more than likely never be false. User registration is never
populated unless we populate it as the user is an interface, and
there is no setter in the lego API.
* Reverted the user parameter in expandDNSChallenge as it is not
used.
* Completed the removal of account_ref which does not seem to be
available in the certificate resource anymore. This includes a state
migration on acme_certificate.
* Removed an unused test fixture.