Add new DNS providers for lego v2.2.0
This adds support for, and documents, all of the DNS providers available
in lego as of v2.2.0.

Another documentation sync will happen to make sure that the rest of the
providers are up to date with the current settings (several poll tuning
options have been added since the docs were initially added).