r/certificate: No partial, dns_challenge is list, no longer ForceNew
I wrote this provider when my understanding of Terraform was not 100%
complete, this fixes some issues that were the result of that:

* Partial state mode was probably never needed and as such has been
removed. No attributes are set in a way that would produce an
inconsistent state mid-flight.
* Complex structures should be lists, not sets. This prevents odd
hash issues from coming up when changing values in blocks that are
especially restricted to just one element only. dns_challenge is one
of these items and as such has been modified to be a list.
* Not necessarily a mistake in schema design, but dns_challenge
configuration items do not need to be ForceNew, as it's perfectly
reasonable that DNS might migrate providers during the lifetime of a
certificate, and since DNS is only used for validating a challenge,
it technically has nothing to do with the rest of the certificate
data. As such this has been removed and one can now change DNS
challenge settings in the middle of a certificate's lifetime without
having to re-create the certificate.

Fixes #1.