r/certificate: fix P12 check, use only bytes decoded
When decoding padded data, DecodedLen will actually return the
maximum *possible bytes* that are needed for a buffer, not the
actual amount of bytes decoded. This means that we need to actual get
the actual amount of bytes decoded from Decode, and slice the buffer
on that length to get the proper data set. If we don't do that,
trailing data will exist on the PKCS12 ASN unmarshal.

This also fixes the decode on the check to use padded standard
decoding.