Add challenge support for multiple DNS providers
Multiple dns_challenge stanzas can now be specified to present the DNS
challenge token. Previously, in situations with multiple primary DNS
providers, the challenge may fail if only one provider was presented
with the challenge token.

The acme_certificate schema was updated to move recursive_nameservers
into the root scope. The schema version was bumped and a migration was
provided.