r/certificate: ensure multiProviders test has a chance to be skipped
The acceptance test framework's skipping logic (skip a test if TF_ACC
is not set) does not run until resource.Test is run, which will
evaluate everything needed to properly pass in the TestCase literal,
including the configuration redering function. This is causing
panics when attempting to reference the data in the nonexistent
ACME_MULTI_PROVIDERS variable.

This adds a small workaround that just provides an empty slice of
length/cap 2 when the variable is not found. This creates an invalid
config, but the config will never be used in this form as the test
will be skipped or failed with this invalid data anyway, as there are
checks on the variable during acceptance testing.