acme: fix migration to schema version 4 (TF 0.11)
The schema migration to schema version 4 has issues in that the
recursive_nameservers key was being migrated "manually", ie: by name,
as if  it was a string value and not a list value. This is not the
case.

This fixes it to the correct behavior.

A schema version bump or state correction should not be necessary:

* A state already migrated to schema version 4 in TF 0.11 will have
had an update effected to it with the correct recursive_nameserver
data overwriting the bad migrated data. Effectively, the state
migration would have been ineffective in facilitating a no-op during
the update, which was the point.
* An upgrade attempt with TF 0.12 would have completely failed, with
Terraform giving an "unknown attribute" error. This is due to the
fact that the resource currently lacks a StateUpgradeFunc.

Effectively this should correct state migrations for those upgrading
from 1.2.1 to 1.3.1 and have no effect on those moving from 1.3.0 to
1.3.1 - on Terraform 0.11.

Terraform 0.12 schema migrations will be added in another commit
before we go to 1.3.1.