acme: decouple and version current resource schemas
This takes the shared schema structure and de-couples it, ensuring
that both resources have their own full copy of the schema without
being at all affected by the other.

It also begins *versioning* the schemas. This is important going
forward to facilitate state migration using the new 0.12 SDK.

The generic resourceACMERegistration and resourceACMECertificate
functions now point to the correct, current, in-use version, with
each specific version of the schema having the version number, ie:
resourceACMERegistrationV1 and resourceACMECertificateV4, V3, etc.
This is then given to a specific updater to get the correct schema
for state migration.

The schema tests have also been removed, which have generally been
low-value on part of the declarative nature of schema definitions,
and are now even more so now that we are using declarative schemas
completely.