r/certificate: get certificate from CA if missing in state
An existing bug with certificate updates means that the
certificate_pem field will be dropped if there are errors renewing.

This does not correct that issue 100% (partial state is needed for
this), but what this *does* do is recover this field by fetching the
certificate data from the CA again if the certificate_pem field is
missing on refresh.

The actual fix to ensure the field is not cleared at all will be
coming in another commit.