r/certificate: use partial mode during certificate renewals
This enables partial mode during the update process when the
certificate needs to be renewed.

This to prevent the resource from writing a bad diff if there is an
error during renewal. Specifically, certificate_pem being changed to
an unknown value during diff customization ultimately results in a
blank value being written if a new value is not set during update,
even in the event of an error.

PR #59 corrects existing states that may have been affected by this
issue by pulling the certificate_pem field directly from the CA.