r/certificate: ensure expired certs get renewed despite HasChange
HasChange currently has some issues working with SetNewComputed, see
hashicorp/terraform#21977. While that gets resolved, the proposed
workaround is to use the same logic (if possible) that's used to
check if SetNewComputed needs to be used as a condition for updating
the resource.

This commit applies this advice to acme_certificate, which is
affected by this issue.

Fixes #73.