modules: update lego to v2.7.1, using modules fork
Note that this is just the update. There are no additional changes
that have been added here, so we still need to add all of the
providers that have been added since 2.5.0.

This PR moves the repository over to using forks for:

* lego, which I have converted to modules
* Terraform, which has an outdated version of the Azure SDK and needs
updating

This should contain the updates to the Google Cloud provider that are
required to stabilize it.