provider: auto-generate supported DNS providers and documentation
This adds the tooling necessary to auto-generate the supported DNS
providers.

Every few releases lego adds a new provider or two and groking the
documentation and new providers every time is a labor. I've wanted to
do this for a while, so decided to give it a go.

The new generator in build-support/generate-dns-providers will look
in the local package cache for the current in-use version of lego,
and pull in the providers, looking in their metadata files (named
PROVIDER_NAME.toml - this is where lego itself sources its docs), and
will then generate the documentation off of this data. It also
generates a provider factory that gets dropped into the acme package.

To get the provider list the generator crawls the providers/dns
directory for the TOML files. This is the same logic lego itself
uses, so it should be stable.