r/registration: remove deactivated or deleted accounts on refresh
This change will remove a deactivated or deleted account from the
state if it's found on refresh. This will allow a destroy plan to
succeed when the account has already been deactivated or is otherwise
missing.

Note that once an account is deactivated, the private key is
unusable, so you cannot use this functionality to delete the resource
and recreate it. You must start a plan with a new private key.