Add pre_check_delay option
This adds the pre_check_delay option, which allows for a delay for DNS
propagation after all DNS challenges have been added.

Propagation delays are an issue on certain DNS providers regardless of
the ability of lego and acme_certificate to monitor NS servers for
success. The error rate is as bad as approximately 1 in every 10.

This option allows for some workaround for problematic providers.

Co-authored-by: Chris Marchesi <chrism@vancluevertech.com>