testing: start of adding support for pebble (WIP)
This is the first effort of adding support for testing with pebble,
LetsEncrypt's own testing implementation of boulder.

  https://github.com/letsencrypt/pebble/

The intention is to eventually use this to test EAB support in the ACME
provider, and possibly longer term move most testing to it as to remove
dependencies on AWS, etc for automated testing (and hence dependencies
on credentials/resources that cost money).

Registration tests were tested manually and are functional. The new
certificate test currently *almost* works, but is stuck on the fact that
I'd like to optimally use the pebble-challtestsrv to do DNS challenges,
as per our long term intent here.

I'm expecting some of the setup/teardown scripts to be modified and the
test be tuned a bit depending on what kinds of challenges we face in
getting this set up just the way that we want. One thing I think might
might be another challenge is that EAB "seems" to be an all-or-nothing
thing in pebble, so two instances would need to be set up. That will
require additional consideration.