r/acme_certificate: add option for disabling propagation check
Following up on f4f286a, doing this disables the propagation check in
lego, allowing us to fully use pebble.

The pebble test is now functional, and next steps will probably entail
cleaning up/disabling the rest of the tests, so that we can just run
pebble out of GitHub actions without the need for credentials to real
infrastructure.