Move provider to standalone provider SDK
This is a intermediary step, we will be moving things to the v2 provider
SDK after this commit.

One aside, we have needed to fork the TLS provider for the time being as
the go.mod file in the repository is currently broken. I will probably
put in a PR after fully upgrading to the v2 SDK.