r/certificate: id is now a UUID
Fixes #103.

This changes the logic so that the ID for the acme_certificate resource
is now a UUID, and not the URL of the certificate, which generally
changes on renewal as a new unique certificate is issued. The resulting
unexpected ID drift is causing issues with downstream resources.

Rather than force a new resource on the certificate expiry pre-check,
which would revoke more than likely currently valid and in-production
certificates, this solution ultimately rightfully directs certificate
URL dependencies to the "certificatet_url" field, which interpolates
properly in these situations as it is updated post-apply.