Fix command line args
This uses exec.Command to populate the initial exec.Cmd instead of doing
it manually, this ensures that os.Args is set properly, and possibly
heads off windows issues at the pass if there were any.

This also takes off the extra plugin name sugar at the end as this was
purely cosmetic and I added it because of the strange issue where
os.Args[0] was not the program name, but that was because of the manual
use of exec.Cmd, and not due to, say, some un-preventable OS issue.