dnsplugin: add timeout support
This was missed in beta1, I want to make sure that we don't lose the
ability to get timeout details. This just moves the existing logic in
the wrapper partly down to the plugin level and makes every returned
client technically a challenge.ProviderTimeout (but you just get zeros
when querying when it doesn't exist - this is the behavior the wrapper
had before this).