Use local version of go-plugin temporarily
This resolves a long-running dependency conflict for the grpc package
that is the result of go-plugin relying on an old version of
protoreflect. This seems to have been resolved in in later versions.
Using my forked version here for now while we wait for an upstream fix.
Depending on the speed of the fix, we may release for now with this pin.