Update flake stuff
I'm moving to a full NixOS setup as a daily driver, so we're migrating
some of the things over there, namely the nix authoring bits which I'm
now managing out of my own workstation config.

Go here is also updated to go_1_21 until the package is fully upstreamed
in nixpkgs.