r/registration: fix "inconsistent result after apply" errors
This fixes an issue where, when destroying a registration and
re-creating it by doing something like renaming the resources,
"inconsistent result after apply" errors may occur.

This was actually a red herring in the fact that it's actually masking
other errors that happen during resource creation, such as the fact that
a registration with the private key exists, but was deactivated
(something that seemingly does not show up during actual creation of an
account at Let's Encrypt, I guess).

The fix here should allow these kinds of errors to properly show up.

Fixes #374.