r/certificate: DNS provider wrapper sequential provider support
This ensures that our DNS provider wrapper supports sequential
providers.

This is an internal interface used by lego to determine if a specific
DNS provider needs to be executed serially (per-domain record) or can be
executed in parallel (all domain records at once). A number of providers
(such as exec and rfc2136) are such providers - there's no guarantee
that parallel operation in these scenarios would be stable, so each
challenge needs to be executed one-by-one.

The delay between executions is controlled on a provider-by-provider
basis, usually through the PROVIDERNAME_SEQUENCE_INTERVAL, where
PROVIDERNAME is the name of the provider.

For multiple providers, the following constraints are enforced:

* Any sequential provider in the set makes the whole set sequential.

* The sequence interval for the set is taking from the provider with the
  longest delay configured.