r/registration: Allow private key to be managed by resource
This sets the private_key_pem attribute in acme_registration to
Optional/Computed, allowing the resource to manage the key fully (and
hence removing the need to have the key supplied by the likes of
tls_private_key).

The goal here is to facilitate easier resource management when working
with EAB attributes - not needing to supply the private key ensures that
you can change the EAB binding, have the old registration deactivated,
and the new one activated with a new private key. Without that last
part, registration will fail, as the unchanged private key would still
be associated with the deactivated registration. If using the
tls_private_key resource to supply the key, one would need to taint that
resource, and with other (out-of-band) methods, the private key would
need to be changed by other means.

Aside from this goal, however, moving the private key to be managed by
the resource markedly simplifies resource use in general, as the need to
supply an external key is now relegated to edge cases, none of which we
have examples for at this time.