r/registration: Add some conflicts to schema
And a basic test for the account_key_pem/account_key_algorithm setting.

The resource CRUD ignores the account key management settings when the
key is supplied, but this just ensures that you don't enable both
without realizing it and then get confused.

Note that there's some deeper things that we're not blocking right now,
such as setting RSA bits when you've got ECDSA selected, but that's
because that requires deeper non-config validation (e.g. CustomizeDiff),
which I think is excessive at this point. NOTE: I'm not too sure if this
is implemented in newer TF APIs, but again, still probably excessive at
this point.