r/certificate: Add profile support
This adds the "profile" attribute to acme_certificate, allowing you to
control the profile used to generate the certificate at the CA.

For example, at Let's Encrypt, this can be used to request a cert
without a common name and shorter authorization validity times
("tlsserver"), or one with a shorter lifetime ("shortlived").

The default is to not request a profile (this is the "classic" profile
at Let's Encrypt).

Fixes #485.