r/certificate: allow the use of ARI (ACME Renewal Information)
This allows the ability to use ARI endpoints (ACME Renewal Information,
RFC 9773) with the acme_certificate resource.

There are two settings (also a debug setting, not shown here but
documented and also can be seen in the commit) that control this:

* The use_renewal_info setting, which controls whether or not ARI data
  is actually used (note that ARI details are always fetched, regardless
  of whether or not they are used).

* The renewal_info_max_sleep setting controls a sleep interval
  (configurable from 0-900 seconds) that allows setting a skew for
  determining whether or not a diff will be able to hit the selected ARI
  renewal time.

The ARI logic works like so: the renewal window for the certificate
is fetched, saved, and a random time is selected within the renewal
window, all in refresh. During diff time, if use_renewal_info is enabled
and the selected renewal time can be reached through a combination of
the current time and renewal_info_max_sleep, a certificate diff is
generated. Renewal then happens normally (with the addition of the
specific fields that need to be set to tell the CA that the order is an
ARI renewal, and is replacing a specific certificate), and the process
resets.

In the event that the selected renewal time cannot be reached (either by
it elapsing or through sleeping), or no ARI endpoint is found, renewal
logic falls back to the more simple min_days_remaining check. This also
prevents soft-locks due to expired certificates, as RFC-conforming ARI
clients are not supposed to query endpoints for expired certificates.
The main thing to note for consumers is that min_days_remaining may need
to be tweaked for certificates with a window that reaches into the
default 30-day min_days_remaining setting.