r/certificate: always deactivate authorizations after cert obtain
This adds a flag (enabled by default) that sets the certificate
creation/renewal process to drop authorizations after a successful
order.

Normally these authorizations linger for quite a while - on Let's
Encrypt, this is 30 days for the classic profile and 7 hours for others.
Presumably, this facilitates scenarios where it's useful for a challenge
to not have to be fulfilled every time a certificate for a particular
identifier is requested. However, from a principle of least privilege,
it makes more sense to make sure that authorization is obtained on an
as-needed basis. For example, this should help prevent instances where
the key is compromised but access to the domain validation process was
missing (e.g., no ability to change DNS records or access to a
well-known HTTP endpoint).